requirements.txt
streamlit
torch
pillow
numpy
requests
